make sky status work with API server
