+++ sky/api/rest.py
@@ -1,4 +1,5 @@
-from typing import Optional
+import pickle
+from typing import Optional, List, Dict, Any
 
 import fastapi
 import pydantic
@@ -9,7 +10,7 @@ from sky import execution
 from sky import optimizer
 from sky.api import sdk
 
-app = fastapi.FastAPI(root_path='/api/v1', debug=True)
+app = fastapi.FastAPI(prefix='/api/v1', debug=True)
 
 
 class LaunchBody(pydantic.BaseModel):
@@ -58,6 +59,57 @@ async def launch(
         _is_launched_by_spot_controller=launch_body.
         _is_launched_by_spot_controller,
     )
+
+class StatusBody(pydantic.BaseModel):
+    cluster_names: Optional[List[str]] = None
+    refresh: bool = False
+
+class StatusReturn(pydantic.BaseModel):
+    name: str
+    launched_at: Optional[int] = None
+    last_use: Optional[str] = None
+    status: Optional[sky.ClusterStatus] = None
+    handle: Optional[dict] = None
+    autostop: int
+    to_down: bool
+    metadata: Dict[str, Any]
+
+
+@app.get('/status')
+async def status(status_body: StatusBody = StatusBody()) -> List[StatusReturn]:
+    clusters = sdk.status(
+        cluster_names=status_body.cluster_names,
+        refresh=status_body.refresh,
+    )
+    status_returns = []
+    for cluster in clusters:
+        status_returns.append(StatusReturn(
+            name=cluster['name'],
+            launched_at=cluster['launched_at'],
+            last_use=cluster['last_use'],
+            handle=cluster['handle'].to_config(),
+            status=cluster['status'],
+            autostop=cluster['autostop'],
+            to_down=cluster['to_down'],
+            metadata=cluster['metadata'],
+        ))
+    return status_returns
+
+
+class DownBody(pydantic.BaseModel):
+    cluster_names: List[str] =  []
+    purge: bool = False
+
+
+@app.post('/down')
+async def down(down_body: DownBody):
+    for cluster_name in down_body.cluster_names:
+        sdk.down(cluster_name=cluster_name, purge=down_body.purge)
+
+@app.get('/health')
+async def health() -> str:
+    return 'OK'
+
 
 
 app.include_router(sdk.app_router)
